comments system and refactoring
