+++ client/src/components/Pitch/Badge/image.jsx
