+++ client/src/components/Pitch/Badge/index.jsx
