+++ client/src/components/Pitch/Badge/props.js
