+++ client/src/components/index.js
@@ -3,11 +3,14 @@ import { lazy } from 'react'
 export const AuthActions = lazy(() => import('components/Auth/Actions'))
 export const AuthLogo = lazy(() => import('components/Auth/Logo'))
 export const Banners = lazy(() => import('components/Banners'))
-export const Comments = lazy(() => import('components/Comments'))
-export const EmailAndPass = lazy(() => import('components/EmailAndPass'))
+export const EmailAndPass = lazy(() => import('components/Auth/EmailAndPass'))
 export const Logout = lazy(() => import('components/Auth/Logout'))
 export const Main = lazy(() => import('components/Main'))
 export const ProfileImage = lazy(() => import('components/ProfileImage'))
 export const ProfilePitches = lazy(() => import('components/Pitches'))
-export const PitchBadge = lazy(() => import('components/PitchBadge'))
+export const PitchBadge = lazy(() => import('components/Pitch/Badge'))
+export const PitchBanner = lazy(() => import('components/Pitch/Banner'))
+export const PitchUser = lazy(() => import('components/Pitch/User'))
+export const PitchActions = lazy(() => import('components/Pitch/Actions'))
+export const PledgesAndComments = lazy(() => import('components/Pitch/PledgesAndComments'))
 export const Searchbox = lazy(() => import('components/Searchbox'))
