+++ client/src/containers/index.jsx
@@ -1,9 +1,9 @@
 import { lazy } from 'react'
 
 export const Ask = lazy(() => import('containers/ask'))
-export const Explore = lazy(() => import('containers/explore'))
 export const Home = lazy(() => import('containers/home'))
 export const Profile = lazy(() => import('containers/profile'))
-export const Question = lazy(() => import('containers/question'))
+export const Pitch = lazy(() => import('containers/pitch'))
+export const Search = lazy(() => import('containers/search'))
 export const Signup = lazy(() => import('containers/signup'))
 export const Signin = lazy(() => import('containers/signin'))
